+++ client/src/components/App/App.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { BrowserRouter } from "react-router-dom";
+import Router from "./Router";
+
+function App() {
+  return (
+    <BrowserRouter basename="/">
+      <Router />
+    </BrowserRouter>
+  );
+}
+
+export default App;
